# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FreshCart API
  description: |
    REST API for the FreshCart online grocery platform.
    
    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `/api/auth/login` or `/api/auth/register`.
    Include it as: `Authorization: Bearer <token>`
    
    ## Roles
    - `customer` — can browse, cart, order, wishlist
    - `admin` — full platform management
    - `delivery_partner` — view assigned orders, update status
  version: 1.0.0
  contact:
    name: FreshCart API Support
    email: api@freshcart.example.com
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://api.freshcart.example.com/api
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /products:
    get:
      summary: List/search/filter products
      tags: [Products]
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (min 2 chars)
        - name: category
          in: query
          schema:
            type: string
          description: Category slug
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating_desc, newest]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated product list
...
  /products/{slug}:
    get:
      summary: Get product detail
      tags: [Products]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product detail
        '404':
          description: Product not found
...
  /orders:
    get:
      summary: Get order history
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated order history
    post:
      summary: Place order
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [addressId, deliverySlotId, paymentMethod]
              properties:
                addressId:
                  type: string
                  format: uuid
                deliverySlotId:
                  type: string
                  format: uuid
                paymentMethod:
                  type: string
                  enum: [upi, card, cod]
      responses:
        '201':
          description: Order placed
        '400':
          description: Validation error or out-of-stock items
...
  /admin/dashboard:
    get:
      summary: Get admin dashboard stats
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard statistics
        '403':
          description: Admin role required
...